pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dealcloser"
version = "0.1.2"
description = "Deal Closer AI – outreach generator"
authors = [{ name = "Luca Pisani" }]
requires-python = ">=3.11"
dependencies = [
  "openai>=1.30.0",
  "typer>=0.12.0",
  "jinja2>=3.1.3",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
  "flask>=3.0.0"
]

[project.scripts]
dealcloser = "dealcloser.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }
